server/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
strawberry-graphql[flask]==0.280.0
psycopg[binary,pool]==3.2.9 
